requirements.txt
django==4.2.5
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary